refactor: split out options parsing, schedule fetching, and version fetching
