+++ core/index.js
@@ -1,21 +1,21 @@
 const { DateTime } = require('luxon');
 const semver = require('semver');
 const parsefiles = require('@nodevu/parsefiles');
+const optionsParser = require('./util/prod/optionsParser')
+const versionFetcher = require('./util/prod/versions')
+const scheduleFetcher = require('./util/prod/schedule')
 
 async function core(options) {
 	// parse our user's options and set up our fetch/DateTime implementations
-	const parsedOptions = await parseOptions(options);
+	const parsedOptions = optionsParser(options);
 
-	const fetch = parsedOptions.fetch;
 	const now = DateTime.fromISO(parsedOptions.now);
-	const nodejsIndexURL = parsedOptions.urls.index;
-	const nodejsScheduleURL = parsedOptions.urls.schedule;
-
+	
 	// collect and configure our data sources
-	const rawVersions = await fetch(nodejsIndexURL);
-	const rawSchedule = await fetch(nodejsScheduleURL);
-	const versions = await rawVersions.json();
-	const schedule = await rawSchedule.json();
+	const versions = await versionFetcher(parsedOptions)
+	const schedule = await scheduleFetcher(parsedOptions)
+
+	// instantiate our data object, to be expanded upon
 	const data = {};
 
 	Object.keys(versions).map(async (version) => {
@@ -215,43 +215,6 @@ async function determineCurrentReleasePhase(now, dates = {}) {
 	}
 
 	return result;
-}
-
-// this function allows us to parse user-passed options.
-//
-// this is particularly useful for tests so we can reduce variables
-// and ensure that our test suite is able to be consistent.
-async function parseOptions(options) {
-	// set up our defaults
-	const parsedOptions = {
-		fetch: globalThis.fetch,
-		now: DateTime.now(),
-		urls: {
-			index: 'https://nodejs.org/dist/index.json',
-			schedule:
-				'https://raw.githubusercontent.com/nodejs/Release/master/schedule.json',
-		},
-	};
-
-	// allow the end-user to replace our fetch implementation with another one they prefer.
-	if (options?.fetch) {
-		parsedOptions.fetch = options.fetch;
-	}
-
-	// allow the end-user to provide a custom DateTime. This is particularly useful for tests.
-	if (options?.now) {
-		parsedOptions.now = options.now;
-	}
-
-	if (options?.urls?.index) {
-		parsedOptions.urls.index = options.urls.index;
-	}
-
-	if (options?.urls?.schedule) {
-		parsedOptions.urls.schedule = options.urls.schedule;
-	}
-
-	return parsedOptions;
 }
 
 // super hacky tool to convert our weird DateTime tooling into a boolean
